Abstracts collision logic into collision.js
