+++ lib/game.js
@@ -7,6 +7,7 @@ var canvasHeight =  canvas.height;
 var Boat = require('./boat');
 var Bullet = require('./bullet');
 var fishies = require('./generate-fish');
+var collision = require('./collision');
 // var keyEvents = require('./key-events');
 
 var bullets = [];
@@ -28,29 +29,17 @@ canvas.addEventListener('keydown', function(key) {
 });
 
 function generateFish(fishies) {
-  fishies.forEach(function (fish) { fish.draw(context).move().reverseDirection(canvasWidth);});
+  fishies.forEach(function (fish) { 
+    fish.draw(context).move()
+        .reverseDirection(canvasWidth);
+  });
 }
 
 function generateBullets(bullets) {
   bullets.forEach(function (bullet) {
-    bullet.draw(context).shoot(canvasHeight);
+    bullet.draw(context)
+          .shoot(canvasHeight);
   });
-}
-
-function xCollision(fish, bullet) {
-  if (fish.x <= bullet.x && (bullet.x + bullet.width) <= (fish.x + fish.width)) {
-    return true; 
-  } else {
-    return false;
-  }
-}
-
-function yCollision(fish, bullet) {
-  if (fish.y <= bullet.y && bullet.y <= (fish.y + fish.height)) {
-    return true;
-  } else {
-    return false;
-  }
 }
 
 function checkForCollision() {
@@ -61,14 +50,6 @@ function checkForCollision() {
       }
     });
   });
-}
-
-function collision(fish, bullet) {
-  if (xCollision(fish, bullet) && yCollision(fish, bullet)) {
-    console.log("COLLISION");
-    console.log(fish);
-    return true;
-  }
 }
 
 function Game() {
